Split resource info container to selector and container (refs #192)
