+++ app/shared/modals/resource-info/ResourceInfoContainer.js
@@ -1,43 +1,9 @@
 import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
-import { createSelector, createStructuredSelector } from 'reselect';
 
 import uiActions from 'actions/uiActions';
 import ResourceInfoModal from './ResourceInfo';
-
-function resourcesSelector(state) {
-  return state.data.resources;
-}
-
-function resourceIdSelector(state) {
-  return state.modals.resourceInfo.resourceId;
-}
-
-const resourceSelector = createSelector(
-  resourcesSelector,
-  resourceIdSelector,
-  (resources, id) => resources[id]
-);
-
-function showSelector(state) {
-  return state.modals.resourceInfo.show;
-}
-
-function unitsSelector(state) {
-  return state.data.units;
-}
-
-const unitSelector = createSelector(
-  unitsSelector,
-  resourceSelector,
-  (units, resource) => resource && units[resource.unit]
-);
-
-const selector = createStructuredSelector({
-  resource: resourceSelector,
-  show: showSelector,
-  unit: unitSelector,
-});
+import selector from './resourceInfoSelector';
 
 UnconnectedResourceInfoContainer.propTypes = {
   hideResourceImages: PropTypes.func.isRequired,
